Allow 'enum' parameter to be passed to super().
Passing the enum through the custructor allows it to be used for OpenAPI
schema definitions, an important use-case for marshmallow. Given it's
the first parameter of the function, the risk of it being referenced by
a client library by name is fairly small, and searching through public
repositories in GitHub doesn't reveal any clients where this would be
problematic.